+++ src/core/constants.js
@@ -4,7 +4,7 @@
  */
 
 // Version
-export const VERSION = '2.0.0';
+export const VERSION = '2.0.1';
 
 // Action identifiers
 export const ACTIONS = {